+++ clumioapi/__init__.py
@@ -2,6 +2,8 @@
 # Copyright 2021. Clumio, Inc.
 #
 
+sdk_version = 'SDK_VERSION'
+
 __all__ = [
     'api_helper',
     'configuration',