+++ clumioapi/controllers/backup_mssql_databases_v1.py
@@ -4,6 +4,7 @@
 
 from clumioapi import api_helper
 from clumioapi import configuration
+from clumioapi import sdk_version
 from clumioapi.controllers import base_controller
 from clumioapi.exceptions import clumio_exception
 from clumioapi.models import create_backup_mssql_database_v1_request
@@ -19,6 +20,12 @@ class BackupMssqlDatabasesV1Controller(base_controller.BaseController):
     def __init__(self, config: configuration.Configuration) -> None:
         super().__init__(config)
         self.config = config
+        self.headers = {
+            'accept': 'application/api.clumio.backup-mssql-databases=v1+json',
+            'x-clumio-organizationalunit-context': self.config.organizational_unit_context,
+            'x-clumio-api-client': 'clumio-python-sdk',
+            'x-clumio-sdk-version': f'clumio-python-sdk:{sdk_version}',
+        }
 
     def list_backup_mssql_databases(
         self, limit: int = None, start: str = None, filter: str = None, embed: str = None
@@ -86,14 +93,9 @@ class BackupMssqlDatabasesV1Controller(base_controller.BaseController):
         _query_parameters = {}
         _query_parameters = {'limit': limit, 'start': start, 'filter': filter, 'embed': embed}
 
-        # Prepare headers
-        _headers = {
-            'accept': 'application/api.clumio.backup-mssql-databases=v1+json',
-            'x-clumio-organizationalunit-context': self.config.organizational_unit_context,
-        }
         # Execute request
         try:
-            resp = self.client.get(_url_path, headers=_headers, params=_query_parameters)
+            resp = self.client.get(_url_path, headers=self.headers, params=_query_parameters)
         except requests.exceptions.HTTPError as http_error:
             errors = self.client.get_error_message(http_error.response)
             raise clumio_exception.ClumioException(
@@ -143,16 +145,11 @@ class BackupMssqlDatabasesV1Controller(base_controller.BaseController):
         _query_parameters = {}
         _query_parameters = {'embed': embed}
 
-        # Prepare headers
-        _headers = {
-            'accept': 'application/api.clumio.backup-mssql-databases=v1+json',
-            'x-clumio-organizationalunit-context': self.config.organizational_unit_context,
-        }
         # Execute request
         try:
             resp = self.client.post(
                 _url_path,
-                headers=_headers,
+                headers=self.headers,
                 params=_query_parameters,
                 json=api_helper.to_dictionary(body),
             )
@@ -186,14 +183,9 @@ class BackupMssqlDatabasesV1Controller(base_controller.BaseController):
         )
         _query_parameters = {}
 
-        # Prepare headers
-        _headers = {
-            'accept': 'application/api.clumio.backup-mssql-databases=v1+json',
-            'x-clumio-organizationalunit-context': self.config.organizational_unit_context,
-        }
         # Execute request
         try:
-            resp = self.client.get(_url_path, headers=_headers, params=_query_parameters)
+            resp = self.client.get(_url_path, headers=self.headers, params=_query_parameters)
         except requests.exceptions.HTTPError as http_error:
             errors = self.client.get_error_message(http_error.response)
             raise clumio_exception.ClumioException(