+++ clumioapi/controllers/restored_aws_ebs_volumes_v1.py
@@ -4,6 +4,7 @@
 
 from clumioapi import api_helper
 from clumioapi import configuration
+from clumioapi import sdk_version
 from clumioapi.controllers import base_controller
 from clumioapi.exceptions import clumio_exception
 from clumioapi.models import restore_aws_ebs_volume_v1_request
@@ -16,6 +17,12 @@ class RestoredAwsEbsVolumesV1Controller(base_controller.BaseController):
     def __init__(self, config: configuration.Configuration) -> None:
         super().__init__(config)
         self.config = config
+        self.headers = {
+            'accept': 'application/api.clumio.restored-aws-ebs-volumes=v1+json',
+            'x-clumio-organizationalunit-context': self.config.organizational_unit_context,
+            'x-clumio-api-client': 'clumio-python-sdk',
+            'x-clumio-sdk-version': f'clumio-python-sdk:{sdk_version}',
+        }
 
     def restore_aws_ebs_volume(
         self, body: restore_aws_ebs_volume_v1_request.RestoreAwsEbsVolumeV1Request = None
@@ -39,16 +46,11 @@ class RestoredAwsEbsVolumesV1Controller(base_controller.BaseController):
 
         _query_parameters = {}
 
-        # Prepare headers
-        _headers = {
-            'accept': 'application/api.clumio.restored-aws-ebs-volumes=v1+json',
-            'x-clumio-organizationalunit-context': self.config.organizational_unit_context,
-        }
         # Execute request
         try:
             resp = self.client.post(
                 _url_path,
-                headers=_headers,
+                headers=self.headers,
                 params=_query_parameters,
                 json=api_helper.to_dictionary(body),
             )